modified tag memory implementation
to determine the appropriate tag for breadcrumb generation, the previous
implementation checked if the referring url contains a tag's url slug.
this worked, but if one tag's url slug were a subset of another tag's
url slug, the implementation could return the wrong tag. i therefore
changed the implemetnation to create a dictionary of query string
values, and check if the "tag" value equals a tag's url slug --> select
that tag.