Added backend capabilities for customized page size
In the initial impelmentation of this blog, pageSize was given as a
parameter to various classes (Pagination view component when calculating
pagination values, blogrepository when retrieving post etc.) -- but was
nevertheless hard-coded into the *calls/constructors* for those classes.
It therefore required modifying many different files to enable
customized page size. A better practice would have been giving the view
model classes a PageSize parameter with a default value, rather than
hard coding in the default value as an integer in various method
calls/constructors. This commit implemented the former setup.  A
subsequent commit will create the html form elements necessary for the
user to modify their pageSize value.