customize page size, display counts for category
- implemented HTML element on account settings page which allows user to
select their default page size
- fixed httpost method model binding so that page size is correctly
incorporated into the view model and passed to blog repository for
submission to database

-implemented a new blogrepository method which returns a
dictionary<string, string> where the key is each category's slug and the
entry is a display friendly representation of the number of posts in
that category. this replaced code in used in the customization view
model (achieving better encapsulation and avoiding redundency), and is
also used to display # of posts for each category on the categories
widget (currently by changing each categories name to the name + (x
posts).