Save posts, image upload tests
This is a large commit, part of which should ideally have been included
in the user customization branch. The main additional functionality
consists of a page of posts which the current user has saved, and
buttons on all posts to save or unsave posts. Repository methods for
saving/unsaving posts, counting save posts, checking if a post has been
saved, and retrieving teh current BlogUser added. Crude logic added to
views to a. determine if toggling of save/unsave button is enabled and
b. determine of a particular post has been saved -> show unsave button
instead of save button.

BlogUser class has a list of posts, representing saved posts.

*** KEY POINT***
with this commit, the all list views (specifically, posttemplate partial
view used in list view) and Post views assume that the ViewBag contains
a boolean value representing whether save/unsave buttons should be
toggled (defaults to just showing save button if not, though it could
simply hide both buttons) -- the boolean value should only be false if
the user is not signed in.

If an action method calls the list view without initializing a
ViewBag.SaveUnsave value, **there will be no compilation error but there
will be an error when the user loads that page, as the view will attempt
to access ViewBag.SaveUnsave and throw an error if it doesnt exist**

subsequent implementations should incorporate all relevant data into
viewmodel and post models, such that Viewbag data is not required.