edited action and pagination methods
if the request is not an ajax request, some action methods now return a
full view rather than a partial view. this ensures that things load
properly if the user has javascript disabled.

if the request is an ajax request, pagination view component returns a
view which tells the user to click an above link to load all posts as a
fresh page. future implementations should allow pagination to work
within ajax pages, but the current implementation does not.