refactored ReadOrUnreadViewComponent with MaxContentLengthToDisplay variable
replaced 2 instances of a magic number (in this case 50, the maximum
number of characters to display in contents before insteading showing a
"view more" link) with one variable. Now contains a
maxContentLengthToDisplay variable which is used in both instances, this
will make it easier to modify and potentially to allow for user
customization in the future.